[Feature] 관리자/주최사/유저 역할 및 권한 나누기 -> 추가 기능 및 예외처리, 페스티벌 게시 요청글 CRUD (#25)
* [Init-Update] 주최사 역할 추가를 위한 파일 생성 및 기능 구현

1. 유저 역할 주최사 추가
2. 회원가입 시 주최사 승인 요청 가능
3. 관리자는 주최사 요청한 목록 조회
4. 관리자는 주최사 요청에 승인

* [Update] 예외처리 추가

1. NotFoundException 지정
2. UnauthorizedException 지정
3. GlobalExceptionHandler에 위의 예외처리 추가

* [Update] 관리자 권한으로 유저 삭제

관리자 권한으로 유저를 삭제하는 기능
관리자 외의 권한으로 접근 불가

* [Update] 페스티벌 작성/수정/삭제 기능 권한에 따른 예외처리 추가

1. 허가된 주최사, 관리자는 페스티벌 작성 후 바로 작성 및 게시 가능
2. 주최사의 경우 본인이 작성한 글만 수정/삭제 가능
3. 관리자의 경우 관리자가 작성한 글 수정, 모든 글에 대한 삭제 가능

* [Update] 페스티벌 게시 요청글 CRUD 구현 및 예외처리

1. 일반 사용자는 페스티벌 게시 요청글 작성 후 관리자에게 게시 요청
2. 일반 사용자는 본인이 작성한 글만 수정/삭제 가능
3. 관리자는 모든 글에 대하여 삭제 가능
4. 관리자는 페스티벌 게시 요청 승인

* [Update] 리뷰, 댓글, 좋아요 권한에 따른 예외처리

1. 주최사, 일반 사용자는 리뷰 및 댓글 작성 가능(관리자 불가)
2. 주최사, 일반 사용자는 본인이 작성한 리뷰 및 댓글만 수정/삭제 가능
3. 관리자는 모든 리뷰 및 댓글에 대하여 삭제 가능

4. 주최사, 일반 사용자는 모든 페스티벌, 리뷰, 댓글에 좋아요 가능(관리자 불가)

* [Update] 관리자 기능 및 페스티벌 게시 요청 기능 일부 코드 수정,

1. 관리자 기능 - 컨트롤러 페스티벌 게시 승인 관련 api 명세 수정
2. 관리자 기능 - 서비스 필요하지 않은 예외처리 삭제
3. 페스티벌 게시 요청 기능 - 컨트롤러 api 명세 오타 수정

* [Update] 관리자 기능 컨트롤러 Swagger Tag 추가

+++ src/main/java/com/sparta/lafesta/festival/service/FestivalServiceImpl.java
@@ -1,5 +1,6 @@
 package com.sparta.lafesta.festival.service;
 
+import com.sparta.lafesta.common.exception.UnauthorizedException;
 import com.sparta.lafesta.festival.dto.FestivalRequestDto;
 import com.sparta.lafesta.festival.dto.FestivalResponseDto;
 import com.sparta.lafesta.festival.entity.Festival;
@@ -28,8 +29,11 @@ public class FestivalServiceImpl implements FestivalService {
     @Override
     @Transactional
     public FestivalResponseDto createFestival(FestivalRequestDto requestDto, User user) {
-        // user 권한 확인 예외처리 추후 추가 작성 예정
-        Festival festival = new Festival(requestDto);
+        // 허가되지 않은 주최사, 일반 사용자 접근 시 예외처리
+        if (user.getRole().getAuthority().equals("ROLE_USER")) {
+            throw new UnauthorizedException("해당 요청에 접근할 수 없습니다.");
+        }
+        Festival festival = new Festival(requestDto, user);
         festivalRepository.save(festival);
         return new FestivalResponseDto(festival);
     }
@@ -54,8 +58,18 @@ public class FestivalServiceImpl implements FestivalService {
     @Override
     @Transactional
     public FestivalResponseDto modifyFestival(Long festivalId, FestivalRequestDto requestDto, User user) {
-        // user 권한 확인 예외처리 추후 추가 작성 예정
         Festival festival = findFestival(festivalId);
+
+        // 주최사는 본인이 작성한 글만 수정 가능
+        if (user.getRole().getAuthority().equals("ROLE_ORGANIZER") && !festival.getUser().getId().equals(user.getId())) {
+            throw new UnauthorizedException("본인이 작성한 글만 수정할 수 있습니다.");
+        }
+
+        // 관리자는 관리자가 작성한 글만 수정 가능
+        if (user.getRole().getAuthority().equals("ROLE_ADMIN") && festival.getUser().getRole().getAuthority().equals("ROLE_ORGANIZER")) {
+            throw new UnauthorizedException("주최사가 작성한 글은 관리자 권한으로 수정할 수 없습니다.");
+        }
+
         festival.modify(requestDto);
         return new FestivalResponseDto(festival);
     }
@@ -64,8 +78,14 @@ public class FestivalServiceImpl implements FestivalService {
     @Override
     @Transactional
     public void deleteFestival(Long festivalId, User user) {
-        // user 권한 확인 예외처리 추후 추가 작성 예정
         Festival festival = findFestival(festivalId);
+
+        // 주최사의 경우 본인이 작성한 글만, 관리자는 모든 글에 대하여 삭제 가능
+        if (!festival.getUser().getId().equals(user.getId())
+                && !user.getRole().getAuthority().equals("ROLE_ADMIN")) {
+            throw new UnauthorizedException("해당 요청에 접근할 수 없습니다.");
+        }
+
         festivalRepository.delete(festival);
     }
 
@@ -73,6 +93,11 @@ public class FestivalServiceImpl implements FestivalService {
     @Override
     @Transactional
     public FestivalResponseDto createFestivalLike(Long festivalId, User user) {
+        // 주최사, 일반 사용자는 좋아요 추가 가능(관리자 불가)
+        if (user.getRole().getAuthority().equals("ROLE_ADMIN")) {
+            throw new UnauthorizedException("좋아요에 관한 권한이 없습니다.");
+        }
+
         Festival festival = findFestival(festivalId);
         // 좋아요를 이미 누른 경우 오류 반환
         if (findFestivalLike(user, festival) != null) {
@@ -87,6 +112,11 @@ public class FestivalServiceImpl implements FestivalService {
     // 페스티벌 좋아요 취소
     @Override
     public FestivalResponseDto deleteFestivalLike(Long festivalId, User user) {
+        // 주최사, 일반 사용자는 좋아요 추가 가능(관리자 불가)
+        if (user.getRole().getAuthority().equals("ROLE_ADMIN")) {
+            throw new UnauthorizedException("좋아요에 관한 권한이 없습니다.");
+        }
+
         FestivalResponseDto response = transactionTemplate.execute(status -> {
             Festival festival = findFestival(festivalId);
             // 좋아요를 누르지 않은 경우 오류 반환
